+++ src/index.js
@@ -1,2 +1,4 @@
 
-export {initializeElement} from "./initialization";
+export * from "./data";
+export * from "./interface";
+export * from "./tools";