+++ src/rendering/render.js
@@ -1,18 +1,31 @@
 
-export function render(data, dimensions, slice) {
+import {assert} from "chai";
 
-  const sliceDataLength = dimensions[1] * dimensions[2];
+import {getLinearVoiTransform} from "./voi";
 
-  const rgbaSliceData = new Uint8ClampedArray(sliceDataLength * 4);
+/**
+ * Renders given imageData into the given Uint8ClampedArray usable as CanvasImageData.data
+ *
+ * @param {OstrichImageData} imageData
+ * @param {number} slice
+ * @param {VoiWindow} voiWindow
+ * @param {Uint8ClampedArray} target
+ */
+export function render(imageData, slice, voiWindow, target) {
+
+  const sliceDataLength = imageData.dimensions[1] * imageData.dimensions[2];
+
+  assert.typeOf(target, Uint8ClampedArray.name);
+  assert.equal(target.length, sliceDataLength * 4);
+
+  const voiTransform = getLinearVoiTransform(voiWindow, 255);
 
   let imageDataIndex = slice * sliceDataLength;
   let canvasDataIndex = 3;
   const lastImageDataIndex = (slice + 1) * sliceDataLength;
 
   while (imageDataIndex < lastImageDataIndex) {
-    rgbaSliceData[canvasDataIndex] = data[imageDataIndex++];
+    target[canvasDataIndex] = voiTransform(imageData.data[imageDataIndex++]);
     canvasDataIndex += 4;
   }
-
-  return rgbaSliceData;
 }